quick: Fixing a codeQL warning about type comparison
Fixing a codeQL warning about type comparison between uint8_t and uint16_t. This won't actually cause an issue right now due to the SERIAL_NUMBER_LENGTH being set to 40 and checked as well, but this will prevent future problems if this were to change for any reason.

+++ src/scsi_helper.c
@@ -1904,7 +1904,7 @@ void copy_Serial_Number( uint8_t *pbuf, char *serialNumber )
     uint16_t snLen = M_BytesTo2ByteValue(pbuf[2], pbuf[3]);
     memcpy(serialNumber, &pbuf[4], M_Min(snLen,SERIAL_NUM_LEN));
     serialNumber[M_Min(snLen,SERIAL_NUM_LEN)] = '\0';
-    for (uint8_t iter = 0; iter < SERIAL_NUM_LEN && iter < snLen; ++iter)
+    for (uint16_t iter = 0; iter < SERIAL_NUM_LEN && iter < snLen; ++iter)
     {
         if (!is_ASCII(serialNumber[iter]) || !isprint(serialNumber[iter]))
         {
